requirements.txt
torch
torchvision
matplotlib
numpy
pillow
scikit-learn
streamlit
jupyter
ipykernel
reportlab
